requirements.txt
asgiref==3.4.1
cffi==1.15.0
cryptography==36.0.1
dj-database-url==0.5.0
dj-rest-auth==2.2.1
Django==3.0.8
django-cors-headers==3.10.1
django-filter==21.1
django-heroku==0.3.1
django-rest-knox==4.1.0
djangorestframework==3.13.1
gunicorn==20.1.0
psycopg2==2.8.6
pycparser==2.21
pytz==2021.3
sqlparse==0.4.2
whitenoise==6.0.0
